Re-Fixing netbeans.tmpl
Netbeans CA_Golem.xls is the build scripts containing Ant Tasks to create files similar to the build.sh does. It handles checking the revision numbering and replaces @@@ with the rev number which it automatically increases by 1. No need for the +1 in the tmpl. Would result in incorrect rev number.