worker_ca_idle.js - Added Fire II and Pangaea pages worker_ca_page.js - Added Pangaea tab to Quests worker_ca_quest.je - Added Pangaea to Quests & added General needed quests for Pangaea & Fire II
Pucker moment because this isn't just me doing a build for Ryco, this is my own additions after some directions from Ryco. For the love of whatever you believe in, if this screws up, someone better know how to tell me to fix it! I'm only on Earth II so I have no way of knowing if this will break anything. Also, there's 2 Siege of the Capital in Fire II (Idiot CA Devs). I read that CAAP had problems because of this, so can't promise there won't be one here as well.
*fingers crossed*